Add Record tests to NetStandard and NetFx
